+++ src/main/java/com/codestates/todo/repository/ToDoRepository.java
@@ -1,10 +1,8 @@
 package com.codestates.todo.repository;
 
 import com.codestates.todo.entity.ToDo;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 
-public interface ToDoRepository extends JpaRepository<ToDo, Long> {
+public interface ToDoRepository extends JpaRepository<ToDo, Integer> {
 }
